Connect Add/Edit Emergency Contact page to API
Use a POST or PUT request to create/update the contact
For the Edit version of the page, the contact id is read from the url parameters. A state parameter prevents the form from rendering while the GET by id request processes. pre-fills the form.
Add a loading state during the POST/PUT request because otherwise the form errors display on the cleared form. On a successful request, the page navigates back to the list view
Rename numbers to contact_numbers for agreement with API parameters